requirements.txt
# FastAPI 및 서버
fastapi==0.104.1
uvicorn[standard]==0.24.0

# 데이터베이스
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# 데이터 처리
pandas==2.1.3
numpy==1.26.2

# HTTP 요청 및 재시도
requests==2.31.0
tenacity==8.2.3

# 환경 변수 관리
python-dotenv==1.0.0

# 유틸리티
pydantic==2.5.0
pydantic-settings==2.1.0
